requirements.txt
alembic==1.8.1
APScheduler==3.9.1.post1
autopep8==2.0.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
flake8==6.0.0
Flask==2.2.2
Flask-Migrate==2.6.0
Flask-Script==2.0.5
Flask-SQLAlchemy==3.0.2
greenlet==2.0.1
idna==3.4
importlib-metadata==5.1.0
importlib-resources==5.10.1
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.1
mccabe==0.7.0
pkg_resources==0.0.0
psycopg2==2.9.5
pycodestyle==2.10.0
pyflakes==3.0.1
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
six==1.16.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.44
tomli==2.0.1
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.13
Werkzeug==2.2.2
zipp==3.11.0
